Address Julie's feedback
