+++ lucene/backward-codecs/src/java/org/apache/lucene/backward_codecs/lucene90/Lucene90HnswGraphBuilder.java
@@ -30,8 +30,10 @@ import org.apache.lucene.util.hnsw.NeighborArray;
 import org.apache.lucene.util.hnsw.NeighborQueue;
 
 /**
- * Builder for HNSW graph. See {@link Lucene90HnswRWGraph} for a gloss on the algorithm and the
+ * Builder for HNSW graph. See {@link Lucene90HnswGraph} for a gloss on the algorithm and the
  * meaning of the hyperparameters.
+ *
+ * <p>This class is preserved here only for tests.
  */
 public final class Lucene90HnswGraphBuilder {
 
@@ -50,7 +52,7 @@ public final class Lucene90HnswGraphBuilder {
   private final RandomAccessVectorValues vectorValues;
   private final SplittableRandom random;
   private final BoundsChecker bound;
-  final Lucene90HnswRWGraph hnsw;
+  final Lucene90HnswGraph hnsw;
 
   private InfoStream infoStream = InfoStream.getDefault();
 
@@ -87,7 +89,7 @@ public final class Lucene90HnswGraphBuilder {
     }
     this.maxConn = maxConn;
     this.beamWidth = beamWidth;
-    this.hnsw = new Lucene90HnswRWGraph(maxConn);
+    this.hnsw = new Lucene90HnswGraph(maxConn);
     bound = BoundsChecker.create(similarityFunction.reversed);
     random = new SplittableRandom(seed);
     scratch = new NeighborArray(Math.max(beamWidth, maxConn + 1));
@@ -101,7 +103,7 @@ public final class Lucene90HnswGraphBuilder {
    * @param vectors the vectors for which to build a nearest neighbors graph. Must be an independet
    *     accessor for the vectors
    */
-  public Lucene90HnswRWGraph build(RandomAccessVectorValues vectors) throws IOException {
+  public Lucene90HnswGraph build(RandomAccessVectorValues vectors) throws IOException {
     if (vectors == vectorValues) {
       throw new IllegalArgumentException(
           "Vectors to build must be independent of the source of vectors provided to HnswGraphBuilder()");
@@ -139,7 +141,7 @@ public final class Lucene90HnswGraphBuilder {
   void addGraphNode(float[] value) throws IOException {
     // We pass 'null' for acceptOrds because there are no deletions while building the graph
     NeighborQueue candidates =
-        Lucene90HnswRWGraph.search(
+        Lucene90HnswGraph.search(
             value, beamWidth, beamWidth, vectorValues, similarityFunction, hnsw, null, random);
 
     int node = hnsw.addNode();
