+++ lucene/backward-codecs/src/java/org/apache/lucene/backward_codecs/lucene90/Lucene90HnswVectorsFormat.java
@@ -80,8 +80,8 @@ public class Lucene90HnswVectorsFormat extends KnnVectorsFormat {
   /** Default number of maximum connections per node */
   public static final int DEFAULT_MAX_CONN = 16;
   /**
-   * Default number of the size the size of the queue maintained while searching and the number of
-   * random entry points to sample during a graph construction.
+   * Default number of the size of the queue maintained while searching and the number of random
+   * entry points to sample during a graph construction.
    */
   public static final int DEFAULT_BEAM_WIDTH = 100;
