+++ lucene/core/src/java/org/apache/lucene/codecs/lucene91/Lucene91HnswVectorsFormat.java
@@ -95,8 +95,7 @@ public final class Lucene91HnswVectorsFormat extends KnnVectorsFormat {
   /** Default number of maximum connections per node */
   public static final int DEFAULT_MAX_CONN = 16;
   /**
-   * Default number of the size the size of the queue maintained while searching during a graph
-   * construction.
+   * Default number of the size of the queue maintained while searching during a graph construction.
    */
   public static final int DEFAULT_BEAM_WIDTH = 100;
